+++ app/src/main/java/nextstep/shoppingcart/component/ShoppingTopBar.kt
@@ -1,31 +1,24 @@
 package nextstep.shoppingcart.component
 
-import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.Spacer
-import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.size
-import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.filled.ShoppingCart
-import androidx.compose.material3.Icon
-import androidx.compose.material3.IconButton
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.text.style.TextAlign
-import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 
 @Composable
 fun ShoppingTopBar(
-    title: String,
-    onClickCart : () -> Unit,
-    modifier: Modifier = Modifier
+    modifier: Modifier = Modifier,
+    title : String = "",
+    titleTextAlign: TextAlign = TextAlign.Center,
+    leftActionButtons: @Composable () -> Unit = {},
+    rightActionButtons: @Composable () -> Unit = {},
 ) {
     Row(
         modifier = modifier
@@ -34,33 +27,14 @@ fun ShoppingTopBar(
         verticalAlignment = Alignment.CenterVertically,
         horizontalArrangement = Arrangement.spacedBy(6.dp)
     ) {
-        Spacer(modifier = Modifier.size(48.dp))
+
+        leftActionButtons()
         Text(
             modifier = Modifier.weight(1f),
             text = title,
             style = MaterialTheme.typography.titleLarge,
-            textAlign = TextAlign.Center
+            textAlign = titleTextAlign
         )
-        IconButton(
-            onClick = onClickCart
-        ) {
-            Icon(
-                modifier = Modifier
-                    .size(20.dp),
-                imageVector = Icons.Default.ShoppingCart,
-                contentDescription = "ShoppingCart"
-            )
-        }
-
+        rightActionButtons()
     }
-}
-
-@Preview(name = "ShoppingTopBar", showBackground = true)
-@Composable
-private fun Preview1() {
-    ShoppingTopBar(
-        modifier = Modifier.fillMaxWidth(),
-        title = "상품 목록",
-        onClickCart = {}
-    )
 }
