+++ app/src/main/java/nextstep/shoppingcart/screen/ShoppingListScreen.kt
@@ -9,14 +9,13 @@ import androidx.compose.foundation.lazy.grid.LazyVerticalGrid
 import androidx.compose.foundation.lazy.grid.items
 import androidx.compose.material3.Scaffold
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.remember
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import nextstep.shoppingcart.R
 import nextstep.shoppingcart.component.ShoppingItem
-import nextstep.shoppingcart.component.ShoppingTopBar
+import nextstep.shoppingcart.component.ShoppingListTopBar
 import nextstep.shoppingcart.model.productList
 
 @Composable
@@ -26,7 +25,7 @@ fun ShoppingListScreen(
     Scaffold(
         modifier = modifier,
         topBar = {
-            ShoppingTopBar(
+            ShoppingListTopBar(
                 title = stringResource(id = R.string.shopping_list_title),
                 onClickCart = {
